+++ packages/cli/lib/ReactWX.js
@@ -1,6 +1,6 @@
 /* eslint-disable */
 /**
- * 运行于微信小程序的React by 司徒正美 Copyright 2019-03-19T09
+ * 运行于微信小程序的React by 司徒正美 Copyright 2019-03-22T11
  * IE9+
  */
 
@@ -1019,6 +1019,7 @@ function useComponent(props) {
     var clazz = registeredComponents[is];
     props.key = this.key != null ? this.key : props['data-instance-uid'] || new Date() - 0;
     delete props.is;
+    clazz.displayName = is;
     if (this.ref !== null) {
         props.ref = this.ref;
     }